pyproject.toml
[project]
name = "desman"
version = "0.1.0"
description = "Local, private AI agent connecting Ollama models to business apps via MCP servers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "fastmcp>=0.9.0",
    "mcp>=1.0.0",
    "ollama>=0.3.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gradio = [
    "gradio>=4.40.0",   # legacy ui/app.py only
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "ui"]

[tool.ruff]
line-length = 100
target-version = "py312"
